backend/requirements.txt
# Django Core
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
psycopg2-binary==2.9.9

# Celery
celery==5.4.0
redis==5.0.7
flower==2.0.1

# AWS
boto3==1.34.144

# Utils
python-dateutil==2.9.0
pytz==2024.1
